refactor: converting into typescript the main NaaS code
