+++ package.json
@@ -2,15 +2,27 @@
   "name": "no-as-service",
   "version": "1.0.0",
   "description": "A lightweight API that returns random rejection or no reasons.",
-  "main": "index.js",
+  "main": "index.ts",
+  "type": "module",
   "scripts": {
-    "start": "node index.js"
+    "dev": "tsx watch index.ts",
+    "build": "tsc && copyfiles reasons.json dist/",
+    "start": "node dist/index.js"
   },
-  "author": "hotheadhacker",
+  "author": "vitao",
   "license": "MIT",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^17.4.2",
     "express": "^4.18.2",
     "express-rate-limit": "^7.0.0"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.19",
+    "@types/express": "^5.0.6",
+    "@types/express-rate-limit": "^5.1.3",
+    "copyfiles": "^2.4.1",
+    "tsx": "^4.22.0",
+    "typescript": "^6.0.3"
   }
 }
